# Optimize cases with long potential simple_keys