--- a/scannerc.go
+++ b/scannerc.go
@@ -626,36 +626,17 @@ func trace(args ...interface{}) func() {
 func yaml_parser_fetch_more_tokens(parser *yaml_parser_t) bool {
 	// While we need more tokens to fetch, do it.
 	for {
-		// Check if we really need to fetch more tokens.
-		need_more_tokens := false
-
-		if parser.tokens_head == len(parser.tokens) {
-			// Queue is empty.
-			need_more_tokens = true
-		} else {
-			// Check for any now stale keys in a required position.
-			if !yaml_parser_stale_simple_keys(parser) {
+		if parser.tokens_head != len(parser.tokens) {
+			// If queue is non-empty, check if any potential simple key may
+			// occupy the head position.
+			head_tok_idx, ok := parser.simple_keys_by_tok[parser.tokens_parsed]
+			if !ok {
+				break
+			} else if valid, ok := yaml_simple_key_is_valid(parser, &parser.simple_keys[head_tok_idx]); !ok {
 				return false
+			} else if !valid {
+				break
 			}
-			// Check if any potential simple key may occupy the head position.
-			for i := len(parser.simple_keys) - 1; i >= 0; i-- {
-				simple_key := &parser.simple_keys[i]
-				if !simple_key.possible {
-					break
-				}
-				if simple_key.token_number < parser.tokens_parsed {
-					break
-				}
-				if simple_key.token_number == parser.tokens_parsed {
-					need_more_tokens = true
-					break
-				}
-			}
-		}
-
-		// We are finished.
-		if !need_more_tokens {
-			break
 		}
 		// Fetch the next token.
 		if !yaml_parser_fetch_next_token(parser) {
@@ -864,20 +845,6 @@ func yaml_simple_key_is_valid(parser *yaml_parser_t, simple_key *yaml_simple_key
 	return true, true
 }
 
-func yaml_parser_stale_simple_keys(parser *yaml_parser_t) bool {
-	// fmt.Println(parser.simple_keys_min_possible_index, ":", len(parser.simple_keys))
-	for i := parser.simple_keys_min_possible_index; i < len(parser.simple_keys); i++ {
-		simple_key := &parser.simple_keys[i]
-		if valid, ok := yaml_simple_key_is_valid(parser, simple_key); !ok {
-			return false
-		} else if !valid {
-			parser.simple_keys_min_possible_index = i + 1
-		}
-
-	}
-	return true
-}
-
 // Check if a simple key may start at the current position and add it if
 // needed.
 func yaml_parser_save_simple_key(parser *yaml_parser_t) bool {
@@ -902,6 +869,7 @@ func yaml_parser_save_simple_key(parser *yaml_parser_t) bool {
 			return false
 		}
 		parser.simple_keys[len(parser.simple_keys)-1] = simple_key
+		parser.simple_keys_by_tok[simple_key.token_number] = len(parser.simple_keys) - 1
 	}
 	return true
 }
@@ -918,9 +886,7 @@ func yaml_parser_remove_simple_key(parser *yaml_parser_t) bool {
 		}
 		// Remove the key from the stack.
 		parser.simple_keys[i].possible = false
-		if parser.simple_keys_min_possible_index > i {
-			parser.simple_keys_min_possible_index = i
-		}
+		delete(parser.simple_keys_by_tok, parser.simple_keys[i].token_number)
 	}
 	return true
 }
@@ -953,10 +919,8 @@ func yaml_parser_decrease_flow_level(parser *yaml_parser_t) bool {
 	if parser.flow_level > 0 {
 		parser.flow_level--
 		last := len(parser.simple_keys) - 1
+		delete(parser.simple_keys_by_tok, parser.simple_keys[last].token_number)
 		parser.simple_keys = parser.simple_keys[:last]
-		if parser.simple_keys_min_possible_index > last {
-			parser.simple_keys_min_possible_index = last
-		}
 	}
 	return true
 }
@@ -1032,6 +996,8 @@ func yaml_parser_fetch_stream_start(parser *yaml_parser_t) bool {
 
 	// Initialize the simple key stack.
 	parser.simple_keys = append(parser.simple_keys, yaml_simple_key_t{})
+
+	parser.simple_keys_by_tok = make(map[int]int)
 
 	// A simple key is allowed at the beginning of the stream.
 	parser.simple_key_allowed = true
@@ -1336,6 +1302,7 @@ func yaml_parser_fetch_value(parser *yaml_parser_t) bool {
 
 		// Remove the simple key.
 		simple_key.possible = false
+		delete(parser.simple_keys_by_tok, simple_key.token_number)
 
 		// A simple key cannot follow another simple key.
 		parser.simple_key_allowed = false